+++ modules/agent-framework/agent-service/src/main/java/org/apache/airavata/agent/connection/service/models/AgentTunnelAck.java
@@ -1,6 +1,7 @@
 package org.apache.airavata.agent.connection.service.models;
 
 public class AgentTunnelAck {
+    private String executionId;
     private int status;
     private String error;
 
@@ -18,5 +19,13 @@ public class AgentTunnelAck {
 
     public void setError(String error) {
         this.error = error;
+    }
+
+    public String getExecutionId() {
+        return executionId;
+    }
+
+    public void setExecutionId(String executionId) {
+        this.executionId = executionId;
     }
 }
