+++ modules/agent-framework/agent-service/src/main/java/org/apache/airavata/agent/connection/service/models/AgentTunnelCreateRequest.java
@@ -1,0 +1,31 @@
+package org.apache.airavata.agent.connection.service.models;
+
+public class AgentTunnelCreateRequest {
+    private String agentId;
+    private int localPort;
+    private String localBindHost;
+
+    public String getAgentId() {
+        return agentId;
+    }
+
+    public void setAgentId(String agentId) {
+        this.agentId = agentId;
+    }
+
+    public int getLocalPort() {
+        return localPort;
+    }
+
+    public void setLocalPort(int localPort) {
+        this.localPort = localPort;
+    }
+
+    public String getLocalBindHost() {
+        return localBindHost;
+    }
+
+    public void setLocalBindHost(String localBindHost) {
+        this.localBindHost = localBindHost;
+    }
+}
