--- modules/agent-framework/agent-service/src/main/java/org/apache/airavata/agent/connection/service/models/AgentTunnelRequest.java
@@ -1,67 +1,0 @@
-package org.apache.airavata.agent.connection.service.models;
-
-public class AgentTunnelRequest {
-    private String agentId;
-    private String destinationHost;
-    private String destinationPort;
-    private String sshUserName;
-    private String sourcePort;
-    private String sshKeyPath;
-    private String password;
-
-    public String getAgentId() {
-        return agentId;
-    }
-
-    public void setAgentId(String agentId) {
-        this.agentId = agentId;
-    }
-
-    public String getDestinationHost() {
-        return destinationHost;
-    }
-
-    public void setDestinationHost(String destinationHost) {
-        this.destinationHost = destinationHost;
-    }
-
-    public String getDestinationPort() {
-        return destinationPort;
-    }
-
-    public void setDestinationPort(String destinationPort) {
-        this.destinationPort = destinationPort;
-    }
-
-    public String getSshUserName() {
-        return sshUserName;
-    }
-
-    public void setSshUserName(String sshUserName) {
-        this.sshUserName = sshUserName;
-    }
-
-    public String getSourcePort() {
-        return sourcePort;
-    }
-
-    public void setSourcePort(String sourcePort) {
-        this.sourcePort = sourcePort;
-    }
-
-    public String getSshKeyPath() {
-        return sshKeyPath;
-    }
-
-    public void setSshKeyPath(String sshKeyPath) {
-        this.sshKeyPath = sshKeyPath;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
