+++ modules/agent-framework/agent-service/src/main/java/org/apache/airavata/agent/connection/service/models/AgentTunnelTerminateResponse.java
@@ -1,0 +1,4 @@
+package org.apache.airavata.agent.connection.service.models;
+
+public class AgentTunnelTerminateResponse {
+}
