+++ modules/agent-framework/airavata-agent/pkg/types.go
@@ -1,0 +1,8 @@
+package pkg
+
+import (
+	protos "airavata-agent/protos"
+	"google.golang.org/grpc"
+)
+
+type Stream = grpc.BidiStreamingClient[protos.AgentMessage, protos.ServerMessage]
